requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-google-genai
google-generativeai
pydantic==1.10.12
python-dotenv